Fix crash bug when NPCs have no bodyparts
Discovered when importing a stock smaug area that had mobs with no bodyparts listed.  Yes, I could have made all the mobs have bodyparts (and most of them SHOULD), but this particular mob was a "collapsing building".   Also noted, while I was here, that bodyparts aren't randomly selected from the list of bodyparts. This needs to be corrected in a future update. - Pars